Refactor facilities_query
